% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allmods.R
\name{allmods}
\alias{allmods}
\title{All Possible Models}
\usage{
allmods(dv="dv", ivs=c("iv1","iv2","iv3"),
        data=dataset, plot = FALSE)
}
\arguments{
\item{dv}{A string of the variable name of you dependent variable.}

\item{ivs}{A list of strings of the names of you independent variables.}

\item{data}{A \code{data.frame} object.}

\item{plot}{A \code{logical} statement indicating whether you want the affiliated
plots.}
}
\value{
A matrix with the R^2, adj R^2, PRESS, and Mallow Cp for all possible
    (non interaction) models given the variables.
}
\description{
This function produces the coeffiecients and the anova summary table
    for Hierarchical regression.
}
\examples{
dv <- rnorm(500, 47, 12)
iv1 <- rnorm(500, 0, 6)
iv2 <- rnorm(500, 100, 5)
iv3 <- rnorm(500, 12, 3)
iv4 <- rnorm(500, 12, 100)
dataset <- data.frame(dv, iv1, iv2, iv3, iv4)
allmods(dv ="dv", ivs=c("iv1", "iv2", "iv3"),
        data = dataset, plots = FALSE)
}
\author{
K. Preston and D. Sambrano
}
