% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cVeda.R
\name{categoriseCvedaReligionLanguageCaste}
\alias{categoriseCvedaReligionLanguageCaste}
\title{Convert the opentext Language Religion Caste variables found in SDIM and ACEIQ 
to the categorical format used in later versions}
\usage{
categoriseCvedaReligionLanguageCaste(df)
}
\arguments{
\item{df}{with the following variables:
language language responses
religion religion responses
caste caste responses
religion_specify language Other Specify Response
language_specify religion Other Specify Response
caste_specify caste Other Specify Responses}
}
\value{
df with categories replaced and uncategorised resposnes shifted to other specify
}
\description{
Convert the opentext Language Religion Caste variables found in SDIM and ACEIQ 
to the categorical format used in later versions
}
