% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Centering Data}
\usage{
center(data)
}
\arguments{
\item{data}{data A matrix, or dataset of class data.frame.}

\item{c.vec}{A vector of center values, defaults to mean of
each variable.}
}
\value{
Returns a dataframe or matrix with all variables centered
}
\description{
Produces the centered data (e.g., new mean = 0 )
}
\examples{
dv <- rnorm(500, 47, 12)
iv1 <- rnorm(500, 0, 6)
iv2 <- rnorm(500, 100, 5)
iv3 <- rnorm(500, 12, 3)
iv4 <- rnorm(500, 12, 100)
dataset <- data.frame(dv, iv1, iv2, iv3, iv4)
center(dataset)
}
\author{
K. Preston and D. Sambrano
}
