% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagen.R
\name{convertFU3toFU2}
\alias{convertFU3toFU2}
\title{Convert FU3 to FU2 data format}
\usage{
convertFU3toFU2(df, subInstrumentSuffix = "FU3", retainAdditionalData = TRUE)
}
\arguments{
\item{df}{Data frame with FU3 data for an instrument or instruments represented by a single task at FU2}

\item{subInstrumentSuffix}{suffix to donate sub instruments (default FU3)}

\item{retainAdditionalData}{whether to retain data outside of subinstruments (default TRUE)}
}
\value{
named list of Data Frames in FU2 format representing the component FU2 instruments contained in the FU3
}
\description{
Convert FU3 variable names and any Value recodes necessary
to return to the format expected from Imagen FU2
}
