% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{decode_base64}
\alias{decode_base64}
\title{Decode a base64 string to a vector of raw bytes}
\usage{
decode_base64(b64)
}
\arguments{
\item{b64}{Single character string containing base64 encoded values}
}
\value{
raw vector
}
\description{
Decode a base64 string to a vector of raw bytes
}
