% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{DelosisAuthenticate}
\alias{DelosisAuthenticate}
\title{Authenticate
local authentication cache prompts for login in shiny window and will cache details for a study and server over repeated calls.}
\usage{
DelosisAuthenticate(SMAusername, studyID, server = "www.delosis.com",
  resetCache = FALSE)
}
\arguments{
\item{SMAusername}{Username to access SMA}

\item{studyID}{Study ID}

\item{server}{SMA server defaults to www.delosis.com}

\item{resetCache}{Force a new prompt defaults to FALSE}
}
\description{
Authenticate
local authentication cache prompts for login in shiny window and will cache details for a study and server over repeated calls.
}
\keyword{authentication}
\keyword{dataset}
\keyword{download}
