% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveAAQ}
\alias{deriveAAQ}
\title{Generate summary for AAQ}
\usage{
deriveAAQ(df)
}
\arguments{
\item{df}{data frame containing long form AAQ data}
}
\value{
wide form of AAQ data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
\details{
NB items A_x and GCP_x are reverse coded in place - they are reversed in the returned df with an R suffix
}
