% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveASSIST}
\alias{deriveASSIST}
\title{Generate summary for ASSIST Questionnaire}
\usage{
deriveASSIST(df)
}
\arguments{
\item{df}{data frame containing long form ASSIST data}
}
\value{
wide form of ASSIST data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
