% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveBART}
\alias{deriveBART}
\title{Derive BART data.}
\usage{
deriveBART(df)
}
\arguments{
\item{df}{Data frame with BART data, read from CSV file exported from Delosis server.}
}
\value{
Derived data frame with summaries.
}
\description{
Applies to the
\href{https://doi.org/10.1037/1076-898X.8.2.75}{Balloon Analogue Risk Task}
implemented in Psytools.
}
