% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveBIG5}
\alias{deriveBIG5}
\title{Generate summary for BIG5}
\usage{
deriveBIG5(df)
}
\arguments{
\item{df}{data frame containing long form BIG5 data}
}
\value{
wide form of BIG5 data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
\details{
NB items 2, 6, 8, 9, 12, 18, 21, 23, 24, 27, 31, 34, 35, 37, 41, 43 are reverse coded in place - they are reversed in the returned df with an R suffix
}
