% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNU.R
\name{deriveBnuBG}
\alias{deriveBnuBG}
\title{Generate summary for BG questionnaire}
\usage{
deriveBnuBG(df)
}
\arguments{
\item{df}{data frame containing long form BG data}
}
\value{
wide form of BG data with summary vars
}
\description{
All this does is to calculate AGE variables from the DOBs provided
}
\details{
NB This does not select the appropriate attempt - this should be done by the calling function
}
