% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNU.R
\name{deriveBnuCOPE}
\alias{deriveBnuCOPE}
\title{Generate summary for COPE questionnaire}
\usage{
deriveBnuCOPE(df)
}
\arguments{
\item{df}{data frame containing long form COPE data}
}
\value{
wide form of COPE data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
