% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNU.R
\name{deriveBnuCWB}
\alias{deriveBnuCWB}
\title{Generate summary for CWB questionnaire}
\usage{
deriveBnuCWB(df)
}
\arguments{
\item{df}{data frame containing long form CWB data}
}
\value{
wide form of CWB data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
