% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNU.R
\name{deriveBnuER}
\alias{deriveBnuER}
\title{Generate summary for ER questionnaire}
\usage{
deriveBnuER(df)
}
\arguments{
\item{df}{data frame containing long form ER data}
}
\value{
wide form of ER data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
