% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNU.R
\name{deriveBnuLS}
\alias{deriveBnuLS}
\title{Generate summary for LS questionnaire}
\usage{
deriveBnuLS(df)
}
\arguments{
\item{df}{data frame containing long form LS data}
}
\value{
wide form of LS data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
