% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNU.R
\name{deriveBnuMS}
\alias{deriveBnuMS}
\title{Generate summary for MS questionnaire}
\usage{
deriveBnuMS(df)
}
\arguments{
\item{df}{data frame containing long form MS data}
}
\value{
wide form of MS data with summary vars
}
\description{
NB the script sent seemed to have an error in the reverse coding of the MX_fixed items
I have assumed that those items should have been fully reversed as the comments seemed to support this idea
}
\details{
NB This does not select the appropriate attempt - this should be done by the calling function
}
