% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNU.R
\name{deriveBnuMW70}
\alias{deriveBnuMW70}
\title{Generate summary for Motivation and wellbeing-70 (MW70) questionnaire}
\usage{
deriveBnuMW70(df)
}
\arguments{
\item{df}{data frame containing long form PDS data}
}
\value{
wide form of PDS data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
