% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNU.R
\name{deriveBnuOCD}
\alias{deriveBnuOCD}
\title{Generate summary for BISBAS questionnaire}
\usage{
deriveBnuOCD(df)
}
\arguments{
\item{df}{data frame containing long form BISBAS data}
}
\value{
wide form of BISBAS data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
