% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNU.R
\name{deriveBnuSC}
\alias{deriveBnuSC}
\title{Generate summary for SC questionnaire}
\usage{
deriveBnuSC(df)
}
\arguments{
\item{df}{data frame containing long form SC data}
}
\value{
wide form of SC data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
