% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNU.R
\name{deriveBnuSimpleSum}
\alias{deriveBnuSimpleSum}
\title{Generate summary for basic questionnaire with summary total}
\usage{
deriveBnuSimpleSum(df, Qname)
}
\arguments{
\item{df}{data frame containing long form data}

\item{Qname}{Name contained in all variables to be summed}
}
\value{
wide form of data with sum
}
\description{
Generate summary for basic questionnaire with summary total
}
