% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNU.R
\name{deriveBnuSPSRQ}
\alias{deriveBnuSPSRQ}
\title{Generate summary for SPSRQ questionnaire}
\usage{
deriveBnuSPSRQ(df)
}
\arguments{
\item{df}{data frame containing long form SPSRQ data}
}
\value{
wide form of SPSRQ data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
