% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveCORSI}
\alias{deriveCORSI}
\title{Derive CORSI data.}
\usage{
deriveCORSI(df)
}
\arguments{
\item{df}{Data frame with CORSI data, read from CSV file exported from Delosis server.}
}
\value{
Derived data frame with summaries.
}
\description{
Applies to the
\href{https://doi.org/10.1207/S15324826AN0704_8}{CORSI block-tapping test}
implemented in Psytools.
}
