% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cVeda.R
\name{deriveCvedaACEIQ}
\alias{deriveCvedaACEIQ}
\title{Generate summary for ACE-IQ questionnaire}
\usage{
deriveCvedaACEIQ(df)
}
\arguments{
\item{df}{data frame containing long form ACE-IQ data}
}
\value{
wide form of ACE-IQ data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
\details{
recodes religion language and caste as required for cVEDA
}
