% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cVeda.R
\name{deriveCvedaAnthropometry}
\alias{deriveCvedaAnthropometry}
\title{Generate summary for cVeda's Anthropometry task - including homogenising units}
\usage{
deriveCvedaAnthropometry(df)
}
\arguments{
\item{df}{data frame containing long form cVeda Antropomtery data}
}
\value{
wide form of Antropomtery data with summary vars and hand entered artefacts removed
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
