% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cVeda.R
\name{deriveCvedaEATQr}
\alias{deriveCvedaEATQr}
\title{Generate summary for EATQ}
\usage{
deriveCvedaEATQr(df)
}
\arguments{
\item{df}{data frame containing long form EATQ data}
}
\value{
wide form of EATQ data with EATQr_SELF_NEGAFFECT_1 corrected
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
\details{
The summary scores are already generated at run time
However EATQr_SELF_NEGAFFECT_1 was incorrectly specified
to include EATQr_35 instead of EATQr_25
}
