% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cVeda.R
\name{deriveCvedaSDIM}
\alias{deriveCvedaSDIM}
\title{Generate summary for SDIM questionnaire}
\usage{
deriveCvedaSDIM(df)
}
\arguments{
\item{df}{data frame containing long form SDIM data}
}
\value{
wide form of SDIM data with language religion and caste recoded
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
\details{
There are no specified summaries for this questionnaire - but the language religion and caste must be recoded
}
