% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveDS}
\alias{deriveDS}
\title{Derive DS data.}
\usage{
deriveDS(df)
}
\arguments{
\item{df}{Data frame with DS data, read from CSV file exported from Delosis server.}
}
\value{
Derived data frame with summaries.
}
\description{
Derive DS data.
}
