% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveERT}
\alias{deriveERT}
\title{Derive ERT data.}
\usage{
deriveERT(df)
}
\arguments{
\item{df}{Data frame with ERT data, read from CSV file exported from Delosis server.}
}
\value{
Derived data frame with summaries.
}
\description{
Applies to the
\href{https://dx.doi.org/10.2466/PMS.104.2.589-598}{Emotion Recognition Task}
implemented in Psytools.
}
