% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveIFVCS}
\alias{deriveIFVCS}
\title{Generate summary for IFVCS}
\usage{
deriveIFVCS(df)
}
\arguments{
\item{df}{data frame containing long form IFVCS data}
}
\value{
wide form of IFVCS data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
