% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagen.R
\name{deriveImagenADRS}
\alias{deriveImagenADRS}
\title{Generate summary for ADRS questionnaire}
\usage{
deriveImagenADRS(df)
}
\arguments{
\item{df}{data frame containing long form ADRS data}
}
\value{
wide form of ADRS data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
