% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagen.R
\name{deriveImagenAUDIT}
\alias{deriveImagenAUDIT}
\title{Generate summary for AUDIT questionnaire}
\usage{
deriveImagenAUDIT(df)
}
\arguments{
\item{df}{data frame containing long form AUDIT data}
}
\value{
wide form of AUDIT data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
\details{
Note that in the case of no alcohol consumption this returns 0 for the summaries
  The original SPSS did not do this but it seems appropriate
  
22.04.2021 - allowed calculation of just freq subscale if administered alone (imacov followup)
}
