% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagen.R
\name{deriveImagenCIS}
\alias{deriveImagenCIS}
\title{Generate summary for CIS questionnaire}
\usage{
deriveImagenCIS(df)
}
\arguments{
\item{df}{data frame containing long form CIS data}
}
\value{
wide form of CIS data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
