% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagen.R
\name{deriveImagenCSI}
\alias{deriveImagenCSI}
\title{Generate summary for CSI questionnaire}
\usage{
deriveImagenCSI(df)
}
\arguments{
\item{df}{data frame containing long form CSI data}
}
\value{
wide form of CSI data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
