% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagen.R
\name{deriveImagenCTS}
\alias{deriveImagenCTS}
\title{Generate summary for CTS questionnaires (from all stages)}
\usage{
deriveImagenCTS(df)
}
\arguments{
\item{df}{data frame containing long form CTS data}
}
\value{
wide form of Reliability data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
\details{
The FU / FU2 and FU3 files use a slightly different structure - this is homogenised here
}
