% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagen.R
\name{deriveImagenIDENT}
\alias{deriveImagenIDENT}
\title{Derive IDENT (Morph) data.}
\usage{
deriveImagenIDENT(df)
}
\arguments{
\item{df}{Data frame with IDENT data, read from CSV file exported from Delosis server.}
}
\value{
Derived data frame with summaries.
}
\description{
Applies to the
Morphed Emotions Task implemented in Psytools for the Imagen Study .
Will be applicable for other Morph type paradigms
}
\details{
Note this function processes all Iterations provided
The calling function should first select the iteration required

This is NOT calculating the thresholds as the original SPSS script did
This could be added in but Frauke did not request them.
}
