% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagen.R
\name{deriveImagenIRI}
\alias{deriveImagenIRI}
\title{Generate summary for IRI questionnaire}
\usage{
deriveImagenIRI(df)
}
\arguments{
\item{df}{data frame containing long form IRI data}
}
\value{
wide form of IRI data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
