% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagen.R
\name{deriveImagenPDS}
\alias{deriveImagenPDS}
\title{Generate summary for PDS questionnaire}
\usage{
deriveImagenPDS(df)
}
\arguments{
\item{df}{data frame containing long form PDS data}
}
\value{
wide form of PDS data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
