% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagen.R
\name{deriveImagenTCI3}
\alias{deriveImagenTCI3}
\title{Generate summary for TCI3 questionnaire}
\usage{
deriveImagenTCI3(df)
}
\arguments{
\item{df}{data frame containing long form TCI3 data}
}
\value{
wide form of TCI3 data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
\details{
The TCI3 had different items in the German FU2 version - for FU3 the German was updated to match the English and French Locales
It seems from discussions with Frauke that actually the German version was the originally intended one and has the most subscales derivable from it
all are merged together for return

The derived varaibles produces depend on which
}
