% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveImgnCIS}
\alias{deriveImgnCIS}
\title{Generate summary for CIS questionnaire}
\usage{
deriveImgnCIS(df)
}
\arguments{
\item{df}{data frame containing long form CIS data}
}
\value{
wide form of CIS data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling funciton
}
