% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagen.R
\name{deriveImgnDOTPROBE}
\alias{deriveImgnDOTPROBE}
\title{Derive DOT_PROBE (Morph) data.}
\usage{
deriveImgnDOTPROBE(df)
}
\arguments{
\item{df}{Data frame with DOT_PROBE data, read from CSV file exported from Delosis server.}
}
\value{
Derived data frame with summaries.
}
\description{
Applies to the
Dot Probe implemented in Psytools for the Imagen Study .
Will be applicable for other dot_probe type paradigms
}
\details{
Note this function processes all Iterations provided
The calling function should first select the iteration required
}
