% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagen.R
\name{deriveImgnGEN}
\alias{deriveImgnGEN}
\title{Generate summary for GENETIC History questionnaire}
\usage{
deriveImgnGEN(df)
}
\arguments{
\item{df}{data frame containing long form GEN data}
}
\value{
wide form of GEN data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
