% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveImgnIRI}
\alias{deriveImgnIRI}
\title{Generate summary for IRI questionnaire}
\usage{
deriveImgnIRI(df)
}
\arguments{
\item{df}{data frame containing long form IRI data}
}
\value{
wide form of IRI data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling funciton
}
\details{
QUERY - in supplied docs Perspective sum contains IRI25 twice?
I have followed this ( ie counting 25 twice) please check
}
