% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveImgnNEO}
\alias{deriveImgnNEO}
\title{Generate summary for NEO FFI questionnaire}
\usage{
deriveImgnNEO(df)
}
\arguments{
\item{df}{data frame containing long form NEO FFI data}
}
\value{
wide form of NEO FFI data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
\details{
Note that in the case of no alcohol consumption this returns 0 for the summaries
  The original SPSS did not do this but it seems appropriate
}
