% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveImgnReliability}
\alias{deriveImgnReliability}
\title{Generate summary for Reliability questionnaires (from all stages)}
\usage{
deriveImgnReliability(df)
}
\arguments{
\item{df}{data frame containing long form Reliability data}
}
\value{
wide form of Reliability data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
\details{
The FU / FU2 and FU3 files use a slightly different structure - this is homogneised here
}
