% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveIPIP20}
\alias{deriveIPIP20}
\title{Generate summary for IPIP 20 item}
\usage{
deriveIPIP20(df)
}
\arguments{
\item{df}{data frame containing long form IPIP data}
}
\value{
wide form of IPIP data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
\details{
NB items 6, 7, 8, 9, 10, 15, 16, 17, 18, 19, 20 are reverse coded in place - they are reversed in the returned df with an R suffix
}
