% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveKIRBY}
\alias{deriveKIRBY}
\title{Derive KIRBY data.}
\usage{
deriveKIRBY(df)
}
\arguments{
\item{df}{Data frame with KIRBY data, read from CSV file exported from Delosis server.}
}
\value{
Derived data frame with summaries AND raw data.
}
\details{
Note this does not select iterations, though it will only produce the Kests for 
   Completed datasets
}
