% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveMID}
\alias{deriveMID}
\title{Derive MID data.}
\usage{
deriveMID(df)
}
\arguments{
\item{df}{Data frame with MID data, read from CSV file exported from Delosis server.}
}
\value{
Data frame with summaries.
}
\description{
Applies to the
\href{https://doi.org/10.1006/nimg.2000.0593}{Monetary Incentive Delay}
task implemented in Psytools.
}
