% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{derivePALP}
\alias{derivePALP}
\title{Derive PALP data.}
\usage{
derivePALP(df)
}
\arguments{
\item{df}{Data frame with PALP data, read from CSV file exported from Delosis server.
If the PALP is administered with separate conditions as separate tasks then they should be merged into a single
long df containing all the different conditions
NB if merging in this way the individual dfs must be limited to a single Iteration per participant beforehand}
}
\value{
Derived data frame with summaries.
}
\description{
Applies to the
\href{https://doi.org/10.1176/appi.ajp.2014.13111499}{PALP task}
implemented in Psytools.
}
