% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveSCQ}
\alias{deriveSCQ}
\title{Generate summary for SCQ questionnaire}
\usage{
deriveSCQ(df)
}
\arguments{
\item{df}{data frame containing long form SCQ data}
}
\value{
wide form of SCQ data with summary vars
}
\description{
Items 5, 9, 13, 17 are reverse coded and returned with R suffix
}
\details{
NB This does not select attempts - this should be done by the calling function
}
