% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveSDQ}
\alias{deriveSDQ}
\title{Generate summary for SDQ questionnaire}
\usage{
deriveSDQ(df)
}
\arguments{
\item{df}{data frame containing long form SDQ data}
}
\value{
wide form of SDQ data with summary vars
}
\description{
Will correct coding if using a 1 based scheme and reverses variables in place adding an R suffix
}
\details{
Works for Parent / Self / Teacher forms generating impact scores if data is supplied

NB This does not select attempts - this should be done by the calling function
}
