% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveSOCRATIS}
\alias{deriveSOCRATIS}
\title{Derive SOCRATIS data.}
\usage{
deriveSOCRATIS(df)
}
\arguments{
\item{df}{Data frame with SOCRATIS data, read from CSV file exported from Delosis server.}
}
\value{
Derived data frame with summaries.
}
\description{
Applies to the
\href{https://doi.org/10.1016/j.ajp.2011.05.014}{Social Cognition Rating Tools in Indian Setting}
implemented in Psytools. This is essentially just a questionnaire -
other questionnaires could be similarly processed!
}
