% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveSRS}
\alias{deriveSRS}
\title{Generate summary for SRS questionnaire}
\usage{
deriveSRS(df)
}
\arguments{
\item{df}{data frame containing long form SRS data}
}
\value{
wide form of SRS data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
