% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveTMT}
\alias{deriveTMT}
\title{Derive TMT data.}
\usage{
deriveTMT(df)
}
\arguments{
\item{df}{Data frame with TMT data, read from CSV file exported from Delosis server.}
}
\value{
Derived data frame with summaries.
}
\details{
Drop the initial practice.
}
