% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveWCST}
\alias{deriveWCST}
\title{Derive WCST data.}
\usage{
deriveWCST(df)
}
\arguments{
\item{df}{Data frame with WCST data, read from CSV file exported from Delosis server.}
}
\value{
Data frame with summaries.
}
\description{
Applies to the
\href{https://doi.org/10.1080/00221309.1948.9918159}{Wisconsin Card Sorting Test}
implemented in Psytools.
}
