% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{downloadSingleDataFile}
\alias{downloadSingleDataFile}
\title{Download Single Data File}
\usage{
downloadSingleDataFile(SMAusername, studyID, taskDigestID,
  server = "www.delosis.com", sampleID = NULL)
}
\arguments{
\item{SMAusername}{username to login with}

\item{taskDigestID}{taskID AND digest ID (eg TASK_ID-DIGEST_ID)}

\item{server}{SMA server defaults to www.delosis.com}

\item{sampleID}{sampleID defaults to NULL}
}
\description{
Download the specified task from the specified server
Authentication cache provided by authenticate function
}
\keyword{dataset}
\keyword{download}
