% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{fixNumericVariables}
\alias{fixNumericVariables}
\title{Coerce numeric vars encoded as character to numeric to facilitate derivations}
\usage{
fixNumericVariables(df)
}
\arguments{
\item{df}{data frame to be converted}
}
\value{
data frame with character encoded numeric variables coerced to numeric
}
\description{
Coerce numeric vars encoded as character to numeric to facilitate derivations
}
