% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierRegression.R
\name{Hier.reg}
\alias{Hier.reg}
\title{Hierarchical Regression}
\usage{
Hier.reg(models=c(fit1, fit2, fit3))
}
\arguments{
\item{model}{Must be a concatinated series of the models to be compared
(e.g., use \code{c()}.}
}
\value{
Two matrices displaying (1) b coeffients at each level of the model
    (2) anova summary table with \eqn{R^2},, adj \eqn{R^2},, delta \eqn{R^2},, delta F-ratios,
    degrees of freedom, and p-values for each step of the model.
}
\description{
This function produces the coeffiecients and the anova summary table
    for Hierarchical regression.
}
\examples{
dv <- rnorm(500, 47, 12)
iv1 <- rnorm(500, 0, 6)
iv2 <- rnorm(500, 100, 5)
iv3 <- rnorm(500, 12, 3)
iv4 <- rnorm(500, 12, 100)
dataset <- data.frame(dv, iv1, iv2, iv3, iv4)
fit1 <- lm(dv ~ iv1, data = dataset)
fit2 <- lm(dv ~ iv1 + iv2, data = dataset)
fit3 <- lm(dv ~ iv1 + iv2 + iv3, data = dataset)
fit4 <- lm(dv ~ iv1 + iv2 + iv3 + iv4, data = dataset)
models <- c(fit1, fit2, fit3, fit4)
Hier.reg(models)
}
\author{
K. Preston and D. Sambrano
}
