% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{labelData}
\alias{labelData}
\title{Label Data Frame from Psytools (Desktop) Questionnaire Resources file}
\usage{
labelData(df, resources)
}
\arguments{
\item{df}{Wide format data frame (or table) to label}

\item{resources}{df containing Psytools resources for this instrument}
}
\description{
Label Data Frame from Psytools (Desktop) Questionnaire Resources file
}
