% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{labelVariable}
\alias{labelVariable}
\title{Label individual row - split out as separate function to simplify loops in labelData}
\usage{
labelVariable(x, Rlabels, Qlabel)
}
\arguments{
\item{x}{vector to label}

\item{Rlabels}{named list to apply to response labels}

\item{Qlabel}{String to apply as question label}
}
\description{
Label individual row - split out as separate function to simplify loops in labelData
}
