% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{MergeAllThatApply}
\alias{MergeAllThatApply}
\title{Convert new style AllThatApply (seperate binary sufixed columns) into old style (single column concatenated with pipe )}
\usage{
MergeAllThatApply(df, grepColumnCollection, finalColumn,
  booleanIndicator = "Y")
}
\arguments{
\item{df}{data table to work on}

\item{grepColumnCollection}{a grepable term to grab all columns that should be merged}

\item{finalColumn}{the new column name to contain the merged data}

\item{booleanIndicator}{- defaults to "Y"}
}
\description{
Convert new style AllThatApply (seperate binary sufixed columns) into old style (single column concatenated with pipe )
}
