% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nplots.R
\name{nplots}
\alias{nplots}
\title{Normality Plots}
\usage{
nplots(data)
}
\arguments{
\item{data}{A vector, matrix, or dataset of class data.frame.}
}
\value{
A 2 by 2 grid of normality plots for each variable in a dataset.
}
\description{
Prints a 2 by 2 grid with normality plots. Plots include: histogram,
    qq-plot (with line), density plot, and box plot. Plots bar graphs
    for nominal data.
}
\examples{
dv <- rnorm(500, 50, 10)
iv1 <- factor(rep(1:2, each = 250))
iv2 <- factor(rep(1:5, 100))
iv3 <- rnorm(500, 12, 3)
iv4 <- rnorm(500, 12, 100)
dataset <- data.frame(dv, iv1, iv2, iv3, iv4)
nplots(dataset)
}
\author{
K. Preston and D. Sambrano
}
