% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{recodeVariables}
\alias{recodeVariables}
\title{Recode select variables in supplied wide df adding "R" suffix to them}
\usage{
recodeVariables(df, varlist, fun)
}
\arguments{
\item{df}{data frame containing wide form data}

\item{varlist}{list of greppable tokens in the df colnames - each must uniquely identify 1 col}

\item{fun}{funtion to apply}
}
\value{
modified df
}
\description{
Recode select variables in supplied wide df adding "R" suffix to them
}
