% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{rotateQuestionnaire}
\alias{rotateQuestionnaire}
\title{Rotate simple questionnaires from long to wide format.}
\usage{
rotateQuestionnaire(df, BlockAsMeasureVar = FALSE, skippedValue = NA,
  idVar = c("User.code", "Iteration", "Language", "Completed",
  "Completed.Timestamp", "Processed.Timestamp"))
}
\arguments{
\item{df}{Data frame with simple questionnaire, read from CSV file exported from Delosis server.}

\item{BlockAsMeasureVar}{Default FALSE - Boolean indicating if the Block should be used as a measureVar when rotating,}

\item{skippedValue}{Default NA - value to apply to omitted questions when rotating}

\item{idVar}{list of variables to identify a row after rotation}
}
\value{
Rotated data frame.
}
\description{
Requires \code{User.code}, \code{Iteration}, \code{Trial} and \code{Trial.result}
columns in input data frame.
Removes repeated occurrences of \code{Trial.result} caused by skipping back.
Should work for any questionnaire to rotate into a wide format, but may want some additional honing!
}
\details{
Preserves the Valid column if it is included in the supplied DF
}
