% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{rotateQuestionnairePreserveBlock}
\alias{rotateQuestionnairePreserveBlock}
\title{Rotate simple questionnaires from long to wide format.
Preserving Block as well as trial for the output}
\usage{
rotateQuestionnairePreserveBlock(df)
}
\arguments{
\item{df}{Data frame with simple questionnaire, read from CSV file exported from Delosis server.}
}
\value{
Rotated data frame.
}
\description{
Requires \code{User.code}, \code{Iteration}, \code{Block},\code{Trial} and \code{Trial.result}
columns in input data frame.
Removes repeated occurrences of \code{Trial.result} caused by skipping back.
Should work for any questionnaire to rotate into a wide format, but may want some additional honing!
}
