% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rowMeansCustomMissing}
\alias{rowMeansCustomMissing}
\title{Utility function to calculate row means first stripping custom missings
and then replacing missing results with a custom missing}
\usage{
rowMeansCustomMissing(
  df,
  customMissingCodes = c(-999, -888, -777, -666),
  missingValue = -666,
  maxMissing = 0
)
}
\arguments{
\item{df}{Data Frame/Table to perform rowSums upon}

\item{customMissingCodes}{list of values to be treated as if they are missing  defaults to c(-999, -888, -777, -666)}

\item{missingValue}{custom missing code to apply to missing results}

\item{maxMissing}{(0 to 1) return a prorated sum if the proportion of missings are under this threshold defaults to 0}
}
\value{
recoded df/dt
}
\description{
Utility function to calculate row means first stripping custom missings
and then replacing missing results with a custom missing
}
