% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sanityCheck}
\alias{sanityCheck}
\title{Check a df meets minimum specs for processing}
\usage{
sanityCheck(df, additionalVars = c(), nonRequiredVars = c())
}
\arguments{
\item{df}{Data frame to be checked.}

\item{additionalVars}{extra columns that must be present in the df}

\item{nonRequiredVars}{standard columns do not need to be present in the df}
}
\value{
boolean
}
\description{
Check a df meets minimum specs for processing
}
