% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleEffects.R
\name{Simple.Effects}
\alias{Simple.Effects}
\title{Simple Effects Analyses}
\usage{
Simple.Effects(model = lm(dv ~ iv1 * iv2,
                                 data = dataset))
}
\arguments{
\item{model}{An object of the lm class.}
}
\value{
A matrix displaying simple effects F-ratios, degrees of freedom, and p-values
}
\description{
This function takes a model of the lm class with two IVs and
produces A matrix with the simple effects F-ratios, degrees of freedom,
and p-values. Currently limited to only 2 factor designs.
}
\examples{
dv <- rnorm(100, 50, 10)
iv1 <- factor(rep(1:2, each = 50))
iv2 <- factor(rep(1:5, 20))
data <- data.frame(dv, iv1, iv2)
model <- lm(dv ~ iv1 * iv2, data)
anova(model)
Simple.Effects(model)
}
\author{
K. Preston and D. Sambrano
}
