% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpSlopes.R
\name{Simple.Slopes}
\alias{Simple.Slopes}
\title{Simple Slopes}
\usage{
Simple.Slopes(model, z.score = FALSE, plot = TRUE,
              plot.colors = c("salmon", "plum4", "steelblue2"), line.width = 2,
              line.type = 1:3, sym = c(15, 17, 19), plot.title = "",
              legend = TRUE, legend.levels = c("high", "middle", "low"),
              legend.location = "topright"))
}
\arguments{
\item{model}{An object of the lm class.}

\item{z.score}{A logical statement indicating whether you want complete centering
(e.g., both mean and sd) or not, default is \code{FALSE}.}

\item{plot}{A logical statement indicating whether you want to produce the
interaction plots, default is \code{TRUE}.}

\item{plot.colors}{A vector of colors of the plot, requires three.}

\item{line.width}{Width of the plot lines.}

\item{line.type}{a vector of the 3 lines produced, requires three values.}

\item{sym}{The symybol at points on the graph. 100 for no symbol.}

\item{plot.title}{The title of the plot}

\item{legend}{a logical vector indicating whether you want a legend for the
plot, default is \code{TRUE}.}

\item{legend.levels}{a vector of the legend labels.}

\item{legend.location}{The location of the legend, default is topright.}
}
\value{
Returns a list of coeffiecients for plus or minus one standard deviation from the mean
    of each independent variable (IV) and displays an plot. Useful for displaying
    interactions of continuous data. The first IV in the model will be plot as the
    x-axis and the second IV will be the different lines.
}
\description{
Produces simple slope coefficients and plots for two predictor models.
    Great for visualizing interactions of continuous data.
}
\examples{
dv <- rnorm(500, 47, 12)
iv1 <- rnorm(500, 0, 6)
iv2 <- rnorm(500, 100, 5)
dataset <- data.frame(dv, iv1, iv2)
model <- lm(dv ~ iv1 * iv2, data = dataset)
Simple.Slopes(model)
}
\author{
K. Preston and D. Sambrano
}
