% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphericity.R
\name{jcsphericity}
\alias{jcsphericity}
\title{Evaluate Commpound Symmetry and Sphericity}
\usage{
jcsphericity(dataset)
}
\arguments{
\item{x}{A matrix or data.frame object.}
}
\description{
Produces a covariance matrix to test for commpound symmetry
    as well as variances for sphericity.
}
\examples{
week1 <- rnorm(10, 0, 1)
week2 <- rnorm(10, 0, 1)
week3 <- rnorm(10, 0, 1)
week4 <- rnorm(10, 0, 1)
week5 <- rnorm(10, 0, 1)
dataset <- data.frame(week1, week2, week3, week4, week5)
jcsphericity(dataset)
}
\author{
K. Preston and D. Sambrano
}
