% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{stripCustomMissings}
\alias{stripCustomMissings}
\title{recode defined missing codes back to NA for Row summing purposes}
\usage{
stripCustomMissings(df, customCodes = c(-999, -888, -777, -666))
}
\arguments{
\item{df}{Data Frame/Table containing custom missing codes}

\item{customCodes}{custom missing codes to code to NA}
}
\value{
recoded df/dt
}
\description{
recode defined missing codes back to NA for Row summing purposes
}
