% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{stripHTML}
\alias{stripHTML}
\title{Utility function to strip html tags from string or vector}
\usage{
stripHTML(htmlString)
}
\arguments{
\item{htmlString}{String /  String Vector to process}
}
\value{
String / Vector with tags removed
}
\description{
Utility function to strip html tags from string or vector
}
