% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ztest.R
\name{z.test}
\alias{z.test}
\title{Z-test}
\usage{
z.test(X, mu, sigma, N)
}
\arguments{
\item{X}{A vector of the data or the mean value of the data.}

\item{mu}{The population mean, default is 0.}

\item{sigma}{The population standard deviation, default is 1.}

\item{N}{Sample Size. Default length of X.}
}
\value{
Produces the standard error of a given input.
}
\description{
This function produces the z statistic of a given input.
}
\examples{
z.test(rnorm(10,3,1))
 #or
z.test(X = 50, mu = 45, sigma = 6, N = 25)
}
\author{
K. Preston and D. Sambrano
}
