% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagen.R
\name{convertFU3toFU2}
\alias{convertFU3toFU2}
\title{Convert FU3 to FU2 data format}
\usage{
convertFU3toFU2(df)
}
\arguments{
\item{df}{Data frame with FU3 data for an instrument or instruments represented by a single task at FU2}
}
\value{
Data Frame in FU2 format.
}
\description{
Convert FU3 variable names and any Value recodes neccessary
to return to the format expected from Imagen FU2
}
