% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNU.R
\name{deriveBnuPDS}
\alias{deriveBnuPDS}
\title{Generate summary for PDS questionnaire as administered to children in BNU}
\usage{
deriveBnuPDS(df)
}
\arguments{
\item{df}{data frame containing long form PDS data}
}
\value{
wide form of PDS data with summary vars
}
\description{
NB - it is NOT suitable for the Adult version as that asks completely differnt questions - no derivation has been defined for that
}
\details{
NB This does not select the appropriate attempt - this should be done by the calling function
}
