% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveCORSI}
\alias{deriveCORSI}
\title{Derive CORSI data.}
\usage{
deriveCORSI(df)
}
\arguments{
\item{df}{Data frame with CORSI data, read from CSV file exported from Delosis server.}
}
\value{
Derived data frame with summaries.
}
\description{
Applies to the
\href{http://digitool.library.mcgill.ca/R/?func=dbin-jump-full&object_id=93903}{CORSI block-tapping test}
(part of the \href{https://dx.doi.org/10.1016/j.jneumeth.2013.10.024}{PEBL} battery)
implemented in Psytools.
}
