% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveDS}
\alias{deriveDS}
\title{Derive DS data.}
\usage{
deriveDS(df)
}
\arguments{
\item{df}{Data frame with DS data, read from CSV file exported from Delosis server.}
}
\value{
Derived data frame with summaries.
}
\description{
Applies to the Digit Span task
(part of the \href{https://dx.doi.org/10.1016/j.jneumeth.2013.10.024}{PEBL} battery)
implemented in Psytools.
}
