% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveImgnESPAD}
\alias{deriveImgnESPAD}
\title{Generate summary for ESPAD questionnaire}
\usage{
deriveImgnESPAD(df)
}
\arguments{
\item{df}{data frame containing long form ESPAD data}
}
\value{
wide form of ESPAD data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
