% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveKIRBY}
\alias{deriveKIRBY}
\title{Derive KIRBY data.}
\usage{
deriveKIRBY(df)
}
\arguments{
\item{df}{Data frame with KIRBY data, read from CSV file exported from Delosis server.}
}
\value{
Derived data frame with summaries AND raw data.
}
\description{
Applies to the
\href{https://www.ncbi.nlm.nih.gov/pubmed/10100392}{Now-or-later test}
(part of the \href{https://dx.doi.org/10.1016/j.jneumeth.2013.10.024}{PEBL} battery))
implemented in Psytools. Again just a questionnaire!
}
\details{
Note this does not select iterations, though it will only produce the Kests for
   Completed datasets
}
