% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveLEQ}
\alias{deriveLEQ}
\title{Generate summary for LEQ item}
\usage{
deriveLEQ(df)
}
\arguments{
\item{df}{data frame containing long form LEQ data - will accept LEQ data with missing items eg BNU / cVEDA}
}
\value{
wide form of LEQ data with summary vars
}
\description{
NB! The _year suffix has had 2 different meanings depending on timepoint
At Imagen baseline _year was a binary yes / no for happened in the last year
At all other timepoints it referred to the AGE at which the event happened
This derivation function will generate _age_mean or _year_meanfreq appropriately
}
\details{
cVEDA follows Imagen baseline

NB This does not select the appropriate attempt - this should be done by the calling function
}
