% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{derivePBI}
\alias{derivePBI}
\title{Generate summary for Alabama Parenting Questionnaire, Child or Parent}
\usage{
derivePBI(df)
}
\arguments{
\item{df}{data frame containing long form APQ data}
}
\value{
wide form of APQ data with summary vars
}
\description{
NB This does not select the appropriate attempt - this should be done by the calling function
}
