% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveSST}
\alias{deriveSST}
\title{Derive SST data.}
\usage{
deriveSST(df)
}
\arguments{
\item{df}{Data frame with SST data, read from CSV file exported from Delosis server.}
}
\value{
Data frame with summaries.
}
\description{
Applies to the
\href{http://dx.doi.org/10.1037/0033-295X.91.3.295}{Stop Signal Task}
task implemented in Psytools.
}
\details{
Clearly there are some people who are simply not responding for big chunks
of this task in c-VEDA. This gives them a good stop hit rate but very bad
go_success rate. They will probably need to be excluded. Really this should
not be happening in a supervised admin situation?
}
