% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{deriveWCST}
\alias{deriveWCST}
\title{Derive WCST data.}
\usage{
deriveWCST(df)
}
\arguments{
\item{df}{Data frame with WCST data, read from CSV file exported from Delosis server.}
}
\value{
Data frame with summaries.
}
