% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psytools_task_derivations.R
\name{selectIteration}
\alias{selectIteration}
\title{Select Iteration from dataset}
\usage{
selectIteration(df, iterationFunction = min, completed = TRUE,
  valid = TRUE)
}
\arguments{
\item{df}{Data frame with simple questionnaire, read from CSV file exported from Delosis server.}

\item{iterationFunction}{function to apply to Iteration - default min}

\item{completed}{restrict to completed Iterations only - default TRUE}

\item{valid}{restrict to Iterations marked as valid IF the user.code has any valid atempts  - default TRUE

It is not clear how the XNAT selection was acheived and the inclusion of invalids should be examined}
}
\value{
data frame complying with input params
}
\description{
Requires \code{User.code}, \code{Iteration}, \code{Completed}
Optionally \code{Valid}
columns in input data frame.
}
\details{
By default returns the df trimmed to the first complete iteration for each user code
}
