% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{selectIteration}
\alias{selectIteration}
\title{Select Iteration from dataset}
\usage{
selectIteration(
  df,
  iterationFunction = min,
  completed = TRUE,
  valid = TRUE,
  allowIncomplete = FALSE,
  allowInvalid = TRUE
)
}
\arguments{
\item{df}{Data frame or table with simple questionnaire, read from CSV file exported from Delosis server.}

\item{iterationFunction}{function to apply to Iteration - default min}

\item{completed}{restrict to completed Iterations only - default TRUE}

\item{valid}{restrict to Iterations marked as valid IF the user.code has any valid attempts - default TRUE}

\item{allowIncomplete}{allow incomplete if a user.code has only incomplete data - default FALSE}

\item{allowInvalid}{allow invalid if a user.code has only invalid data - default TRUE
It is not clear how the XNAT selection was achieved and the inclusion of invalids should be examined}
}
\value{
data frame complying with input params
}
\description{
Requires \code{User.code}, \code{Iteration}, \code{Completed}
Optionally \code{Valid}
columns in input data frame.
}
\details{
By default returns the df trimmed to the first complete iteration for each user code
}
